Create common class for parsing json schema.
