+++ parsers/recipe.py
@@ -1,8 +1,12 @@
+import json
 import requests
 from bs4 import BeautifulSoup
 
 
 class Recipe(object):
+
+    def __init__(self, domain):
+        pass
 
     def fetch_html(self, url):
         #fd = open('allrecipes3.html', 'r')
@@ -13,5 +17,44 @@ class Recipe(object):
 
     def fetch_soup(self, url):
         html = self.fetch_html(url)
-        soup = BeautifulSoup(html)
+        soup = BeautifulSoup(html, features='lxml')
         return soup
+
+class WpJsonRecipe(Recipe):
+    """Some wordpress sites provide the recipe in a convenient json format."""
+
+    def __init__(self, domain):
+        self.domain = domain
+
+    def fetch_json(self, url):
+        soup = self.fetch_soup(url)
+        result = soup.find('script', {'type': 'application/ld+json'})
+        return self.get_json_recipe(json.loads(result.contents[0]))
+
+    def get_json_recipe(self, d):
+        recipe = {}
+        for r in d['@graph']:
+            if not isinstance(r['@type'], str):
+                continue
+
+            if r['@type'].lower() != 'recipe':
+                continue
+
+            recipe['name'] = r['name']
+            recipe['description'] = r['description']
+            recipe['ingredients'] = r['recipeIngredient']
+            recipe['instructions'] = [i['text'] for i in r['recipeInstructions']]
+            recipe['image'] = r['image'][0]
+
+        return recipe
+
+    def Parse(self, url):
+        recipe = {
+                'url': url,
+                'source': self.domain
+                }
+
+        parsed_recipe = self.fetch_json(url)
+        recipe.update(parsed_recipe)
+
+        return recipe
